+++ integration-tests/internal-full-stack.sh
@@ -368,8 +368,8 @@ http {
   ratelimitly_fail     close;
   ratelimitly_debug    on;
 
-  ratelimitly_zone allow_zone bucket="v1|fixture=allow" rate=10000r/s;
-  ratelimitly_zone deny_zone  bucket="v1|fixture=deny"  rate=1r/s;
+  ratelimitly_zone allow_zone "bucket=v1|fixture=allow" rate=10000r/s;
+  ratelimitly_zone deny_zone  "bucket=v1|fixture=deny"  rate=1r/s;
 
   server {
     listen ${NGINX_HOST}:${NGINX_PORT};